+++ projects/brochure-marketplace/src/app/services/api.service.ts
@@ -16,8 +16,6 @@ export abstract class ApiService {
 
   abstract getRegistryPackages(registryUrl: string): Promise<GetPackagesRes>
 
-  abstract getKnownRegistries(): Promise<T.KnownRegistry[]>
-
   abstract getStaticProxy(
     pkg: MarketplacePkg,
     path: 'LICENSE.md' | 'instructions.md',