+++ projects/start-os/web/ui/src/app/services/api/embassy-live-api.service.ts
@@ -354,13 +354,6 @@ export class LiveApiService extends ApiService {
     })
   }
 
-  async getKnownRegistries(): Promise<T.KnownRegistry[]> {
-    return this.rpcRequest({
-      method: 'marketplace.known-registries',
-      params: {},
-    })
-  }
-
   // notification
 
   async getNotifications(