+++ projects/start-os/web/ui/src/app/services/marketplace.service.ts
@@ -4,7 +4,6 @@ import {
   GetPackageRes,
   Marketplace,
   MarketplacePkg,
-  resolveIdentity,
   StoreDataWithUrl,
   StoreIdentity,
 } from '@start9labs/marketplace'
@@ -53,28 +52,25 @@ export class MarketplaceService extends AbstractMarketplaceService {
   private readonly storage = inject(StorageService)
   private readonly i18n = inject(i18nPipe)
 
-  readonly knownRegistries$: Observable<T.KnownRegistry[]> = from(
-    this.api.getKnownRegistries(),
-  ).pipe(
-    catchError(() => of<T.KnownRegistry[]>([])),
-    shareReplay(1),
-  )
+  readonly registries$: Observable<StoreIdentity[]> = this.patch
+    .watch$('ui', 'registries')
+    .pipe(
+      map(registries => [
+        toStoreIdentity(start9, registries[start9]),
+        toStoreIdentity(community, registries[community]),
+        ...Object.entries(registries)
+          .filter(([u, _]) => !sameUrl(start9, u) && !sameUrl(community, u))
+          .map(([url, name]) => toStoreIdentity(url, name)),
+      ]),
+    )
 
-  readonly registries$: Observable<StoreIdentity[]> = combineLatest([
-    this.patch.watch$('ui', 'registries'),
-    this.knownRegistries$.pipe(startWith<T.KnownRegistry[]>([])),
-  ]).pipe(
-    map(([registries, known]) => [
-      resolveIdentity(start9, registries[start9] || null, known),
-      resolveIdentity(community, registries[community] || null, known),
-      ...Object.entries(registries)
-        .filter(([u, _]) => !sameUrl(start9, u) && !sameUrl(community, u))
-        .map(([url, name]) => resolveIdentity(url, name, known)),
-    ]),
-  )
-
-  readonly newRegistry$ = this.registries$.pipe(
-    startWith<StoreIdentity[]>([]),
+  readonly newRegistry$ = this.patch.watch$('ui', 'registries').pipe(
+    map(registries =>
+      [start9, community, ...Object.keys(registries)]
+        .filter((url, i, all) => all.findIndex(u => sameUrl(u, url)) === i)
+        .map(url => ({ url, name: registries[url] ?? null })),
+    ),
+    startWith<{ url: string; name: string | null }[]>([]),
     pairwise(),
     mergeMap(([p, c]) => c.filter(a => !p.find(b => sameUrl(a.url, b.url)))),
   )
@@ -336,5 +332,12 @@ export class MarketplaceService extends AbstractMarketplaceService {
     if (oldName !== newName) {
       this.api.setDbValue<string>(['registries', url], newName)
     }
+  }
+}
+
+function toStoreIdentity(url: string, name?: string | null): StoreIdentity {
+  return {
+    url,
+    name: name || url,
   }
 }